+++ code/CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [#1041](https://github.com/InditexTech/weavejs/issues/1041) Custom guides support and refactor of snapping and distance guides plugins
+- 
+## [3.8.4] - 2026-05-12
+
+### Fixed
+
+- [#1055](https://github.com/InditexTech/weavejs/issues/1055) When adding images with the Images tool, they are not positioned on a grid
 
 ## [3.8.3] - 2026-05-12
 
@@ -1788,7 +1794,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#24](https://github.com/InditexTech/weavejs/issues/24) Bug when loading rooms with text or images
 - [#18](https://github.com/InditexTech/weavejs/issues/18) Fix awareness not working on store-azure-web-pubsub
 
-[Unreleased]: https://github.com/InditexTech/weavejs/compare/3.8.3...HEAD
+[Unreleased]: https://github.com/InditexTech/weavejs/compare/3.8.4...HEAD
+
+[3.8.4]: https://github.com/InditexTech/weavejs/compare/3.8.3...3.8.4
 
 [3.8.3]: https://github.com/InditexTech/weavejs/compare/3.8.2...3.8.3